* bleeding-jumbo:
  Duration limit tweaks to some OpenCL formats.
  OpenCL formats: Use size_t instead of int in a bunch of places because of problems where eg. a GWS of 32M x buffer size of 64 ends up in 2G which doesn't fit in a signed int.

+++ src/opencl_mysqlsha1_fmt.c
@@ -69,7 +69,7 @@ static struct fmt_tests tests[] = {
 	{NULL}
 };
 
-static void create_clobj(int gws, struct fmt_main *self)
+static void create_clobj(size_t gws, struct fmt_main *self)
 {
 	global_work_size = gws;
 	self->params.max_keys_per_crypt = gws;
@@ -122,7 +122,7 @@ static void done(void)
 	HANDLE_CLERROR(clReleaseProgram(program[ocl_gpu_id]), "Release Program");
 }
 
-static cl_ulong gws_test(int gws, int do_benchmark, struct fmt_main *self)
+static cl_ulong gws_test(size_t gws, int do_benchmark, struct fmt_main *self)
 {
 	cl_ulong startTime, endTime;
 	cl_event Event[4];
@@ -175,7 +175,7 @@ static cl_ulong gws_test(int gws, int do_benchmark, struct fmt_main *self)
 	if (do_benchmark)
 		fprintf(stderr, "crypt kernel %.2f ms, ", (double)((endTime - startTime)/1000000.));
 
-	/* 200 ms duration limit for GCN to avoid ASIC hangs */
+	/* 200 ms duration limit */
 	if (endTime - startTime > 200000000) {
 		if (do_benchmark)
 			fprintf(stderr, "- exceeds 200 ms\n");
